Add git big-objects
This command will list large objects in your git repository. By default
you get the 10 largest, but you can choose more or less by providing an
optional extra argument.

I intend to turn this in to a ruby gem to make it easier to install, and
to make it possible to use Thor for the CLI.