Add git squish
Squish repacks a repo to aggressively clean up its contents.

So the equivalent of "git gc --aggressive" - but done *properly* - is to
do (overnight) something like

    git repack -a -d --depth=250 --window=250

where that depth thing is just about how deep the delta chains can be
(make them longer for old history - it's worth the space overhead), and
the window thing is about how big an object window we want each delta
candidate to scan.

And here, you might well want to add the "-f" flag (which is the "drop
all old deltas", since you now are actually trying to make sure that
this one actually finds good candidates.