Stop signing git commits
Unfortunately signing commits with GPG from inside Vim is impractical
when the key is password protected, without storing the password in
insecure way (i.e. in a plain text file).

As I don't consider this an option and want to do manage git from within
Vim more in the future, I will no longer sign commits outside of Vim for
the sake of consistency, and to avoid possible confusion.

I will continue to sign tags outside of Vim for the foreseeable future.