Revert "Add reattach-to-user-namespace in tmux"
This reverts commit c4066b5a9fbc43e3d5d0ef7b591321fa4b91a7b3.

When zsh starts up via reattach-to-user-namespace, rbenv shims is at the
end of $path, meaning the wrong Ruby version gets picked up.

I can reproduce this outside of reattach-to-user-namespace with `zsh -`,
which says to me the problem likely isn't with RTUN, but more likely
something related to rbenv or my dotfiles configuration.

As I'm not desperate to use pbcopy/pbpaste right now this can wait.